split up actions and sagas
