+++ src/actions/user.js
@@ -1,0 +1,10 @@
+import {
+  SET_CURRENT_USER,
+} from '../constants/actionTypes';
+
+const setCurrentUser = user => ({
+  type: SET_CURRENT_USER,
+  payload: { user },
+});
+
+export default setCurrentUser;
